feat: refact infra and add supertest
